j ai oublie l indentation, ca va faire beaucoup de norme a coriger

+++ libft/ft_atoi.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
 	int			i;
 	long int	n;
